Assignment #3 Submission
* Re-targeted build for newer Visual Studio version

* Created basic class for sprites in the game

* Moved window width and height to separate header file

* Created sprite sheet

* Modified GameEntity class to handle UV wrapping and draw calls

* Added the loading of the sprite sheet as a texture in main()

* SetSprite() now maintains the center coordinate of the sprite

Before, whenever the sprite was changed, it would jump to the center of
the screen.

* Created PlayerLaserCannon sprite class

* Created Bullet sprite class

* Created Invader sprite class

* Made MoveX and MoveY public in the Entity class

* Created basic mechanism for moving and shooting the laser cannon

This will serve as the start screen later, where the user will shoot the
Start button to start the game.

* Added title and instruction text to start screen

Next step: detect when a bullet hits the Start or Quit button

* Changed start screen vertices to be use constants

* Swapped order of || operation

More bullets will be traveling up than traveling down, so it makes sense
to check this first.

* Programmed Start and Quit buttons on start screen to respond to bullets

* Changed case of enums

* Switched to using remove_if to delete off-screen bullets

* Switched from a list to a forward_list for bullets

There really is no need to iterate backward through the list.

* Created ASCII character sprite sheet

* Made the SetBox methods in Entity public and used them for the start screen

* Switched to range-based for loop for bullets

* Created SetBox shorthand method

* Added score display

The user can tell that the initial score is 0, I guess.

* Added initial position to Entity constructor

The sprite's center is moved to this position.

* Fixed scale factor in text baseline positioning

* Added indicator for number of lives left

* Moved gameplay variables together

* Moved DrawText to separate file

* Created single function for drawing triangles with textures

* Added player and bullets to gameplay

* Made bullets disappear when hitting the Start button on the start screen

* Ensured that characters are always in range

In DrawText(), any character that is outside of 0x20 and 0xff will be
replaced with a space.

* Added NumberLeft() to the Lives class

This is necessary for other code to know how many lives are left.

* Added initial position parameters to Invader class

* Created MoveOffScreen method for the Bullet class

* Added drawing of invaders during gameplay

Now, we just have to make them move and respond to bullets.

* Flipped rows and columns for list of invaders

* Made invaders react to getting hit by bullets

* Changed invaders to a list of vectors

* Enabled invaders to move

If the invaders get past the player, then the game is over. If the
player shoots all of the invaders, then the game is also over.

* Moved the X position of the column into a register

* Made the difficulty curve quadratic

* Added time step to reduce CPU usage

* Moved input processing to a struct

* Enabled pressing Esc key to return to start screen

* Allowed invaders to shoot back at the player

* Made player lose a life when hit by bullet

* Switched to modelview matrices for transformations

* Fixed loss of precision

There were some double constants that should have been float constants.

* Added entrance and exit animations to start screen

* Made invaders fly in from the right

* Fixed crash when two bullets hit invader

If two bullets simultaneously hit an invader, then the code tries to pop
an invader off the vector twice. If there was only one invader in the
column to begin with, then this would cause a crash.

* Enabled continuous gameplay

When the player successfully destroys a fleet of invaders, another fleet
flys in. The next fleet approaches the player more quickly. The player
gets an additional life after accumulating 1,000 points.

* Added screen for when the player dies

* Stretched start screen by one pixel

This seemed to help with the crispness of the texture.

* Moved ShaderProgram object to global variable

* Moved textures to global variables

* Moved generic rectangle functions to separate class

* Added explosions for bullet strikes during gameplay

* Added explosions for bullet strikes on start screen

* Fixed potentially uninitialized pointer compilation error

* Created simple batch file to launch game

This is for Windows only, obviously.

* Added immortal mode

The player can't die if the string "immortal" is passed in as the first
argument.

* Changed invader entrace to horizontal movement only

* Fixed animations not showing on the start screen